Rename the known_hosts ; disable host hashes
Rename our 'known_hosts' file to the same prefix as the default file,
and with a '_github' suffix.

Host hashing in 'known_hosts' brings no additional security:
- the file is clearly identified as being to identify Github hosts
- the hosts list is already in clear in the 'config' file
- it makes harder for the user to check than only Github hosts are there
So we now disable hashing for hosts identified by their names, but we
will still hash the hosts identified by IP that ssh will automatically
add when connecting to an unknown Github IP: we are not interested in
tracking Github's IPs.