config: if 'man' is missing, look for the man page in well known places
When filtering for options, we extract options from the man page.
To locate the man page we use 'man -w ssh_config'
But on msysgit from bash, the man page is available, but not the 'man'
command. So we look for the file in well known locations.

Note: unfortunately this doesn't solve anything on msysgit because the
ssh_config.5 file we find is not in the man format. :(