Preserve the position of the GHKG section in ~/.ssh/config
Previously the GHKG section was always moved at the end of the file. But
the order counts in ssh_config, so this was breaking configs were a
"Host *" section in present at the end of the file (to define default
settings): the section was moved before GHKG, so it would break our
config.

The parsing of ~/.ssh/config is completely reworked to:
- perserve the position of an existing GHKG section.
- inject a new GHKG section at the end of the file, except if the last
  section is "Host *" (in that case, put it just above)
- recognize a "Host" section in lower case